Crate jf_rescue

This module implements Rescue hash function over the following fields

bls12_377 base field

ed_on_bls12_377 base field

ed_on_bls12_381 base field

ed_on_bn254 base field

It also has place holders for

bls12_381 base field

bn254 base field

bw6_761 base field

Those three place holders should never be used.

Modules

errors

Error types.

Structs

PRP

Rescue pseudo-random permutation (PRP) instance

Permutation

Instance of a unkeyed cryptographic permutation to be used for instantiation hashing, pseudo-random function, and other cryptographic primitives

RescueMatrix

A matrix that consists of STATE_SIZE number of rescue vectors.

RescueVector

Data type for rescue prp inputs, keys and internal data

Constants

RATE

The rate of rescue hash.

ROUNDS

The # of rounds of rescue hash.

STATE_SIZE

The state size of rescue hash.

Traits

RescueParameter

This trait defines constants that are used for rescue hash functions.