Trait jf_rescue::gadgets::RescueGadget

pub trait RescueGadget<R, T, F>
where
    R: SpongeStateVar<T, F>,
{
    // Required methods
    fn rescue_permutation(&mut self, input_var: R) -> Result<R, CircuitError>;
    fn prp(&mut self, key_var: &R, input_var: &R) -> Result<R, CircuitError>;
    fn rescue_sponge_no_padding(
        &mut self,
        data_vars: &[R::Var],
        num_output: usize
    ) -> Result<Vec<R::Var>, CircuitError>;
    fn rescue_sponge_with_padding(
        &mut self,
        data_vars: &[R::Var],
        num_output: usize
    ) -> Result<Vec<R::Var>, CircuitError>;
    fn rescue_full_state_keyed_sponge_no_padding(
        &mut self,
        key: R::Var,
        data_vars: &[R::Var]
    ) -> Result<R::Var, CircuitError>;
    fn rescue_full_state_keyed_sponge_with_zero_padding(
        &mut self,
        key: R::Var,
        data_vars: &[R::Var]
    ) -> Result<R::Var, CircuitError>;
    fn key_schedule(
        &mut self,
        mds_states: &RescueMatrix<R::Native>,
        key_var: &R,
        prp_instance: &PRP<R::Native>
    ) -> Result<Vec<R>, CircuitError>;
    fn create_rescue_state_variable(
        &mut self,
        state: &RescueVector<R::Native>
    ) -> Result<R, CircuitError>;
    fn prp_with_round_keys(
        &mut self,
        input_var: &R,
        mds: &RescueMatrix<R::Native>,
        keys_vars: &[R]
    ) -> Result<R, CircuitError>;
}

Trait for rescue circuit over native field.

Required Methods

fn rescue_permutation(&mut self, input_var: R) -> Result<R, CircuitError>

Given an input state st_0 and an output state st_1, ensure that st_1 = rescue_permutation(st_0) where rescue_permutation is the instance of the Rescue permutation defined by its respective constants

input_var - variables corresponding to the input state

returns - variables corresponding to the output state

fn prp(&mut self, key_var: &R, input_var: &R) -> Result<R, CircuitError>

Rescue based Pseudo Random Permutation (PRP)

key_var - rescue state variable corresponding to the cipher key

input_var - rescue state variable corresponding to the plaintext

returns - state variable corresponding to the cipher text

fn rescue_sponge_no_padding( &mut self, data_vars: &[R::Var], num_output: usize ) -> Result<Vec<R::Var>, CircuitError>

Sponge-based hashes from Rescue permutations

data_vars - sponge input variables, data_vars.len() should be a positive integer that is a multiple of the sponge rate (i.e. 3)

num_output - number of output variables

returns - a vector of variables that refers to the sponge hash output

fn rescue_sponge_with_padding( &mut self, data_vars: &[R::Var], num_output: usize ) -> Result<Vec<R::Var>, CircuitError>

Sponge-based hashes from Rescue permutations

data_vars - sponge input variables,

num_output - number of output variables

returns - a vector of variables that refers to the sponge hash output

fn rescue_full_state_keyed_sponge_no_padding( &mut self, key: R::Var, data_vars: &[R::Var] ) -> Result<R::Var, CircuitError>

Full-State-Keyed-Sponge with a single output

key - key variable

input - input variables,

returns a variable that refers to the output

fn rescue_full_state_keyed_sponge_with_zero_padding( &mut self, key: R::Var, data_vars: &[R::Var] ) -> Result<R::Var, CircuitError>

Similar to Self::rescue_full_state_keyed_sponge_no_padding except data_var are padded with “zero_var”

fn key_schedule( &mut self, mds_states: &RescueMatrix<R::Native>, key_var: &R, prp_instance: &PRP<R::Native> ) -> Result<Vec<R>, CircuitError>

Return the round keys variables for the Rescue block cipher

mds_states - Rescue MDS matrix

key_var - state variable representing the cipher key

returns - state variables corresponding to the scheduled keys

fn create_rescue_state_variable( &mut self, state: &RescueVector<R::Native> ) -> Result<R, CircuitError>

Create a variable representing a rescue state

state - Rescue state

returns - state variables corresponding to the state

fn prp_with_round_keys( &mut self, input_var: &R, mds: &RescueMatrix<R::Native>, keys_vars: &[R] ) -> Result<R, CircuitError>

Return the variable corresponding to the output of the of the Rescue PRP where the rounds keys have already been computed “dynamically”

input_var - variable corresponding to the plain text

mds_states - Rescue MDS matrix

key_vars - variables corresponding to the scheduled keys

returns -