jf_rescue::gadgets::commitment

Trait CommitmentGadget

pub trait CommitmentGadget {
    // Required method
    fn commit(
        &mut self,
        input: &[Variable],
        blinding: Variable,
    ) -> Result<Variable, CircuitError>;
}

Commitment gadget

Required Methods

fn commit( &mut self, input: &[Variable], blinding: Variable, ) -> Result<Variable, CircuitError>

Commitment function.

input - input variables,

blinding - blinding variable

returns a variable that refers to the commitment value The underlying the commitment instance is bound to a specific length. Hence input length must match it.